keep irrelevant codes consistant between different labs to make the changes look incremental

+++ os/lab2/lib/kdebug.c
@@ -25,28 +25,33 @@ int kprintf(const char *fmt, ...)
     va_list ap;
     uint64_t val;
     uint64_t temp;
-    int len;
-    int rev = 0;
+    uint64_t len;
+    uint64_t rev = 0;
     int ch;
     const char *str = NULL;
 
     va_start(ap, fmt);
-    while (*fmt != '\0') {
-        switch (*fmt) {
+    while (*fmt != '\0')
+    {
+        switch (*fmt)
+        {
         case '%':
             fmt++;
-            switch (*fmt) {
-            case 'u': //Decimal
+            switch (*fmt)
+            {
+            case 'u':        //Decimal
                 val = va_arg(ap, uint64_t);
                 temp = val;
                 len = 0;
-                do {
+                do
+                {
                     len++;
                     temp /= 10;
                 } while (temp);
                 rev += len;
                 temp = val;
-                while (len) {
+                while (len)
+                {
                     ch = temp / kpow(10, len - 1);
                     temp %= kpow(10, len - 1);
                     kputchar(ch + '0');
@@ -60,32 +65,38 @@ int kprintf(const char *fmt, ...)
                 val = va_arg(ap, uint64_t);
                 temp = val;
                 len = 0;
-                do {
+                do
+                {
                     len++;
                     temp /= 16;
                 } while (temp);
                 rev += len;
                 temp = val;
-                while (len) {
+                while (len)
+                {
                     ch = temp / kpow(16, len - 1);
                     temp %= kpow(16, len - 1);
-                    if (ch <= 9) {
+                    if (ch <= 9)
+                    {
                         kputchar(ch + '0');
-                    } else {
+                    }
+                    else
+                    {
                         kputchar(ch - 10 + 'a');
                     }
                     len--;
                 }
                 break;
             case 's':
-                str = va_arg(ap, const char *);
+                str = va_arg(ap, const char*);
 
-                while (*str) {
+                while (*str)
+                {
                     kputchar(*str);
                     str++;
                 }
                 break;
-            case 'c': //character
+            case 'c':        //character
                 kputchar(va_arg(ap, int));
                 rev += 1;
                 break;
